+++ tests/PdfBorderTest.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace fpdf;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfBorder::class)]
+#[CoversClass(PdfBorder::class)]
 class PdfBorderTest extends TestCase
 {
     public function testAll(): void