+++ tests/PdfCompareBorderTest.php
@@ -12,9 +12,11 @@ declare(strict_types=1);
 
 namespace fpdf;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(FPDF::class)]
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfDocument::class)]
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfBorder::class)]
+use PHPUnit\Framework\Attributes\CoversClass;
+
+#[CoversClass(FPDF::class)]
+#[CoversClass(PdfDocument::class)]
+#[CoversClass(PdfBorder::class)]
 class PdfCompareBorderTest extends AbstractCompareTestCase
 {
     /**