+++ tests/PdfJpgParserTest.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace fpdf;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfJpgParser::class)]
+#[CoversClass(PdfJpgParser::class)]
 class PdfJpgParserTest extends TestCase
 {
     public function testInvalid(): void