+++ tests/PdfPageSizeTest.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace fpdf;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfPageSize::class)]
+#[CoversClass(PdfPageSize::class)]
 class PdfPageSizeTest extends TestCase
 {
     public function testA4(): void