+++ tests/PdfPointTest.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace fpdf;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfPoint::class)]
+#[CoversClass(PdfPoint::class)]
 class PdfPointTest extends TestCase
 {
     public function testAsArray(): void