+++ tests/PdfRotationTest.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace fpdf;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
-#[\PHPUnit\Framework\Attributes\CoversClass(PdfRotation::class)]
+#[CoversClass(PdfRotation::class)]
 class PdfRotationTest extends TestCase
 {
     public function testValues(): void